feat: add required interfaces for managing kernels
Experience: an a group of kernel commits with a single set of
hyperparameters and a single model architecture. If you modify the
hyperparams or the architecture of the model, then you need to define a
new experience.

Cycle: a cycle is an iterative process than can be divided on many
commits if needed. For example, training a network on 1000 epochs is a
cycle of 1000 steps, it can be divided in sub-cycles if the total time
of execution excedes 9hours, the maximum execution time for a kaggle
kernel.